Machine: fix types in instruction test
Ubuntu18 has problem with expanding within qCompare function.

+++ src/machine/instruction.test.cpp
@@ -22,8 +22,8 @@ void TestInstruction::instruction_access() {
     QCOMPARE(i.rt(), (uint8_t)0x1f);
     QCOMPARE(i.rd(), (uint8_t)0x1f);
     QCOMPARE(i.shamt(), (uint8_t)0x1f);
-    QCOMPARE(i.funct(), (uint8_t)0x3f);
-    QCOMPARE(i.immediate(), (uint16_t)0xffff);
+    QCOMPARE(i.funct(), (uint16_t)0x3f);
+    QCOMPARE(i.immediate(), (int32_t)0xffff);
     QCOMPARE(i.address().get_raw(), (uint64_t)0x3ffffff);
 }
